Refactor currency formatters, use thousands on index page
