+++ src/components/totalAmountItem.js
@@ -1,51 +1,41 @@
 import React from "react"
+import { formatDollars } from "../common/util/formatters"
 import styles from "./totalAmountItem.module.scss"
-
-const currencyFormatter = Intl.NumberFormat("en-US", {
-  style: "currency",
-  currency: "USD",
-  minimumFractionDigits: 0,
-  maximumFractionDigits: 0,
-})
-
-function formatDollars(value) {
-  return currencyFormatter.format(value)
-}
 
 function title(type) {
   switch (type) {
     case "expenditures":
-      return "Total spent";
+      return "Total spent"
     case "balance":
-      return "Current balance";
+      return "Current balance"
     case "measureSupport":
-      return "Total Raised to Support";
+      return "Total Raised to Support"
     case "measureOppose":
-      return "Total Raised to Oppose";
+      return "Total Raised to Oppose"
     default:
-      return "Total raised";
+      return "Total raised"
   }
 }
 
 function detail(type) {
-  let detail;
+  let detail
 
   switch (type) {
     case "expenditures":
-      detail = "Expenditures";
-      break;
+      detail = "Expenditures"
+      break
     case "balance":
-      detail = "Cash on hand";
-      break;
+      detail = "Cash on hand"
+      break
     case "measureSupport":
     case "measureOppose":
-      break;
+      break
     default:
-      detail = "Contributions";
-      break;
+      detail = "Contributions"
+      break
   }
 
-  return detail ? " (" + detail + ")" : "";
+  return detail ? " (" + detail + ")" : ""
 }
 
 export default function TotalAmountItem({ total, type = "contributions" }) {
